+++ iiot_gateway/inc/iiot_gateway.h
@@ -4,7 +4,7 @@
 
 #include "../../common/mqtt/inc/mqtt_utils.h"
 #include "../../common/config/inc/conf_utils.h"
-
+#include "../../common/utils/inc/utils.h"
 
 int core_Init();
 int core_routine();